requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.18.0
requests>=2.31.0
beautifulsoup4>=4.12.3
feedparser>=6.0.10
python-dotenv>=1.0.1
pydantic-settings>=2.1.0
apscheduler>=3.10.4
aiohttp>=3.9.3
Pillow>=10.2.0
gspread>=6.0.0
google-auth>=2.27.0
